[fix] DP helpers are in 'utils' to prevent child-proc. import errors.
