+++ src/vllm_tt_plugin/platform.py
@@ -1,12 +1,10 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: 2025 Tenstorrent USA, Inc.
 
-import ast
 import json
 import multiprocessing
 import os
 import sys
-from contextlib import suppress
 from typing import TYPE_CHECKING, ClassVar, Literal
 
 import torch
@@ -20,6 +18,11 @@ from vllm_tt_plugin.config import (
     validate_tt_lane_config,
 )
 from vllm_tt_plugin.logger import init_tt_logger
+from vllm_tt_plugin.utils.dp_discovery import (
+    StandardDPAssignmentT,
+    _run_standard_dp_visible_device_group_discovery,
+    _split_standard_dp_discovery_result,
+)
 
 if TYPE_CHECKING:
     from vllm.config import VllmConfig
@@ -35,8 +38,6 @@ logger = init_tt_logger(__name__)
 _STANDARD_DP_DISCOVERY_RECV_TIMEOUT_S = 60.0
 _STANDARD_DP_DISCOVERY_JOIN_TIMEOUT_S = 5.0
 _STANDARD_DP_MESH_GRIDS_KEY = "_tt_standard_dp_mesh_grids"
-
-StandardDPAssignmentT = tuple[str, tuple[int, int]]
 
 TT_SCHEDULER_CLS = "vllm_tt_plugin.scheduler.TTScheduler"
 TT_LANE_SCHEDULER_CLS = "vllm_tt_plugin.lane_scheduler.TTLaneCoordinator"
@@ -64,6 +65,7 @@ def _galaxy_generator_version() -> str | None:
 
 
 def _uses_explicit_tt_mpi_launch(vllm_config: "VllmConfig") -> bool:
+    """Returns whether TT must use explicit MPI-based placement."""
     tt_config = get_tt_config(vllm_config)
     parallel_config = vllm_config.parallel_config
     return bool(
@@ -74,94 +76,11 @@ def _uses_explicit_tt_mpi_launch(vllm_config: "VllmConfig") -> bool:
     )
 
 
-_MESH_GRID_PRESETS = {
-    "N150": (1, 1),
-    "P100": (1, 1),
-    "P150": (1, 1),
-    "P150x2": (1, 2),
-    "N300": (1, 2),
-    "P300": (1, 2),
-    "N150x4": (1, 4),
-    "P150x4": (1, 4),
-    "T3K": (1, 8),
-    "P150x8": (1, 8),
-    "P300x2": (1, 4),
-}
-
-
-def _parse_mesh_grid(
-    mesh_device_env: str | None,
-    num_devices_available: int,
-    *,
-    tg_mesh_grid: tuple[int, int],
-) -> tuple[int, int]:
-    mesh_grid_dict = dict(_MESH_GRID_PRESETS)
-    mesh_grid_dict["TG"] = tg_mesh_grid
-
-    if mesh_device_env is None:
-        return (1, num_devices_available)
-
-    try:
-        parsed_value = ast.literal_eval(mesh_device_env)
-        if isinstance(parsed_value, (tuple, list)) and len(parsed_value) == 2:
-            return tuple(int(dim) for dim in parsed_value)
-        raise ValueError("Not a valid tuple")
-
-    except (ValueError, SyntaxError, TypeError):
-        mesh_grid = mesh_grid_dict.get(mesh_device_env)
-        if mesh_grid is None:
-            raise ValueError(
-                f"Invalid MESH_DEVICE: {mesh_device_env}. "
-                f"Expected one of: {list(mesh_grid_dict.keys())}"
-            ) from None
-        return mesh_grid
-
-
-def _resolve_parent_mesh_grid(
-    mesh_device_env: str | None,
-    num_devices_available: int,
-) -> tuple[int, int]:
-    mesh_grid = _parse_mesh_grid(
-        mesh_device_env,
-        num_devices_available,
-        tg_mesh_grid=(4, 8),
-    )
-
-    if mesh_grid[0] * mesh_grid[1] != num_devices_available:
-        mesh_grid = (1, num_devices_available)
-
-    return mesh_grid
-
-
-def _maybe_reorder_standard_dp_visible_device_groups(
-    device_groups: list[StandardDPAssignmentT],
-    mesh_grid: tuple[int, int],
-    data_parallel_size: int,
-) -> list[StandardDPAssignmentT]:
-    import ttnn
-
-    if (
-        ttnn.cluster.get_cluster_type() == ttnn.cluster.ClusterType.GALAXY
-        and mesh_grid == (4, 8)
-        and data_parallel_size == 4
-        and len(device_groups) == 4
-    ):
-        reordered_groups = [device_groups[index] for index in (0, 2, 3, 1)]
-        logger.info(
-            "Reordered TT single-host DP device groups for WH Galaxy DP=4 "
-            "from row-major %s to mesh-id order %s",
-            [visible_devices for visible_devices, _shape in device_groups],
-            [visible_devices for visible_devices, _shape in reordered_groups],
-        )
-        return reordered_groups
-
-    return device_groups
-
-
 def _store_standard_dp_mesh_grids(
     vllm_config: "VllmConfig",
     mesh_grids: dict[str, tuple[int, int]],
 ) -> None:
+    """Store discovered mesh-grid hints on the vLLM config."""
     additional_config = getattr(vllm_config, "additional_config", None)
     if not isinstance(additional_config, dict):
         additional_config = {}
@@ -175,6 +94,7 @@ def _store_standard_dp_mesh_grids(
 def _load_standard_dp_mesh_grids(
     vllm_config: "VllmConfig",
 ) -> dict[str, tuple[int, int]]:
+    """Load stored mesh-grid hints from the vLLM config."""
     additional_config = getattr(vllm_config, "additional_config", None) or {}
     if not isinstance(additional_config, dict):
         return {}
@@ -197,104 +117,8 @@ def _load_standard_dp_mesh_grids(
     return mesh_grids
 
 
-def _split_standard_dp_discovery_result(
-    discovery_result: list[str] | list[StandardDPAssignmentT] | None,
-) -> tuple[list[str] | None, dict[str, tuple[int, int]]]:
-    if discovery_result is None:
-        return None, {}
-    if not discovery_result:
-        return [], {}
-
-    first_entry = discovery_result[0]
-    if isinstance(first_entry, str):
-        return discovery_result, {}
-
-    assignments = discovery_result
-    return (
-        [visible_devices for visible_devices, _mesh_grid in assignments],
-        {visible_devices: mesh_grid for visible_devices, mesh_grid in assignments},
-    )
-
-
-def _discover_standard_dp_visible_device_groups(
-    mesh_device_env: str | None,
-    data_parallel_size: int,
-) -> list[StandardDPAssignmentT]:
-    import ttnn
-    from models.tt_transformers.tt.generator import create_submeshes
-
-    mesh_device = None
-    submeshes = []
-
-    try:
-        num_devices_available = ttnn.get_num_devices()
-        mesh_grid = _resolve_parent_mesh_grid(mesh_device_env, num_devices_available)
-        mesh_device = ttnn.open_mesh_device(ttnn.MeshShape(*mesh_grid))
-        submeshes = create_submeshes(mesh_device, data_parallel_size)
-        if len(submeshes) != data_parallel_size:
-            raise RuntimeError(
-                "TT create_submeshes returned "
-                f"{len(submeshes)} groups for data_parallel_size={data_parallel_size}"
-            )
-
-        device_groups = []
-        for dp_rank, submesh in enumerate(submeshes):
-            device_ids = list(submesh.get_device_ids())
-            if not device_ids:
-                raise RuntimeError(f"TT DP rank {dp_rank} resolved to an empty submesh")
-            device_groups.append(
-                (
-                    ",".join(str(device_id) for device_id in device_ids),
-                    tuple(int(dim) for dim in submesh.shape),
-                )
-            )
-
-        device_groups = _maybe_reorder_standard_dp_visible_device_groups(
-            device_groups,
-            mesh_grid,
-            data_parallel_size,
-        )
-
-        logger.info(
-            "Resolved TT single-host DP device groups: %s",
-            [
-                f"{visible_devices}@{mesh_shape}"
-                for visible_devices, mesh_shape in device_groups
-            ],
-        )
-
-        return device_groups
-
-    finally:
-        for submesh in submeshes:
-            with suppress(Exception):
-                ttnn.close_mesh_device(submesh)
-        if mesh_device is not None:
-            with suppress(Exception):
-                ttnn.close_mesh_device(mesh_device)
-
-
-def _run_standard_dp_visible_device_group_discovery(
-    conn,
-    mesh_device_env: str | None,
-    data_parallel_size: int,
-) -> None:
-    try:
-        conn.send(
-            (
-                "ok",
-                _discover_standard_dp_visible_device_groups(
-                    mesh_device_env, data_parallel_size
-                ),
-            )
-        )
-    except Exception as exc:
-        conn.send(("error", f"{type(exc).__name__}: {exc}"))
-    finally:
-        conn.close()
-
-
 def _terminate_discovery_process(proc: multiprocessing.Process) -> None:
+    """Terminates a lingering discovery subprocess."""
     proc.terminate()
     proc.join(timeout=_STANDARD_DP_DISCOVERY_JOIN_TIMEOUT_S)
     if proc.is_alive():
@@ -306,6 +130,7 @@ def _join_discovery_process_or_raise(
     proc: multiprocessing.Process,
     message: str,
 ) -> None:
+    """Joins a discovery subprocess or raises if it stays alive."""
     proc.join(timeout=_STANDARD_DP_DISCOVERY_JOIN_TIMEOUT_S)
     if proc.is_alive():
         _terminate_discovery_process(proc)
@@ -315,6 +140,18 @@ def _join_discovery_process_or_raise(
 def _resolve_standard_dp_visible_device_groups(
     vllm_config: "VllmConfig",
 ) -> list[StandardDPAssignmentT] | None:
+    """Resolves single-host TT device groups for standard DP.
+
+    Notes
+    -----
+    The discovery work runs in a spawned helper so the parent avoids holding TT
+    chip locks before worker startup.
+
+    Examples
+    --------
+    >>> _resolve_standard_dp_visible_device_groups(vllm_config)
+    [("0,1,2,3", (1, 4)), ...]
+    """
     parallel_config = vllm_config.parallel_config
     if parallel_config.data_parallel_size <= 1 or _uses_explicit_tt_mpi_launch(
         vllm_config
