+++ src/vllm_tt_plugin/utils/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: 2025 Tenstorrent USA, Inc.
+
+"""vLLM Tenstorrent plugin configuration utilities."""
